+++ packages/config-tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # config-tsconfig
 
+## 0.0.3
+
+### Patch Changes
+
+- Standardize message roles
+
 ## 0.0.2
 
 ### Patch Changes