+++ packages/convostack-agent-sse/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @convostack/agent-sse
 
+## 0.0.5
+
+### Patch Changes
+
+- Standardize message roles
+- Updated dependencies
+  - @convostack/shared@0.0.5
+  - @convostack/agent@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes