+++ packages/convostack-langchain-memory/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @convostack/langchain-memory
 
+## 0.0.6
+
+### Patch Changes
+
+- Standardize message roles
+- Updated dependencies
+  - @convostack/shared@0.0.5
+
 ## 0.0.5
 
 ### Patch Changes