+++ packages/schema-graphql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # schema-graphql
 
+## 0.0.3
+
+### Patch Changes
+
+- Standardize message roles
+
 ## 0.0.2
 
 ### Patch Changes